chore(shared): Consolidate tab components
