+++ sites/shared/components/mdx/index.mjs
@@ -3,7 +3,7 @@ import { Highlight } from './highlight.mjs'
 import { YouTube } from './youtube.mjs'
 //import { Figure } from './figure.mjs'
 import { ReadMore } from './read-more.mjs'
-import { Tab, Tabs } from './tabs.mjs'
+import { Tab, Tabs } from '../tabs.mjs'
 import { TabbedExample as Example } from './tabbed-example.mjs'
 import { HttpMethod, HttpStatusCode } from './http.mjs'
 import { ControlTip } from '../control/tip.mjs'
